Tab border adjustments
